fix: add fallback for PagerViewAdapter on web
